# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "clipos API",
    "description": "Invite-only clip archive.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/clips/{id}/analyse": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Read a clip's killfeed again, e.g. after a fix to the analysis. The new result replaces\nthe old one when the worker is done; until then the analysis is pending.",
        "operationId": "analyse_clip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Clip id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Analysis queued, or already queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisQueued"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such clip",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The clip isn't published, or has no killfeed to read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/invites": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Every invite, active first, newest first.",
        "operationId": "list_invites",
        "responses": {
          "200": {
            "description": "Invites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invite"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Invite an email (or re-activate a revoked invite).",
        "operationId": "create_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Invite created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/invites/revoke": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke an invite. The user and their clips stay, but they can no longer sign in.",
        "operationId": "revoke_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Invite revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invite"
                }
              }
            }
          },
          "400": {
            "description": "Revoking your own invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No active invite for that email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Every user who has signed in.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/users/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change a user's role or status.",
        "operationId": "update_user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserAccess"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Changing your own access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "The signed-in user, created on first call.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Not invited, account disabled, or token lacks email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Edit your own profile.",
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "No token, or an invalid or expired one",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Handle taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisQueued": {
        "type": "object",
        "required": [
          "pending"
        ],
        "properties": {
          "pending": {
            "type": "boolean",
            "description": "Always true: the clip's analysis is queued or running."
          }
        }
      },
...
      "CreateInvite": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Role",
                "description": "Defaults to `member`."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Error payload for every non-2xx `/api` response.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code: `bad_request`, `unauthorized`, `not_invited`,\n`account_disabled`, `forbidden`, `not_found`, `conflict`, `unsupported_media_type`,\n`too_large`, `bad_gateway` or `internal`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Never contains internals for 5xx responses."
          },
          "tied": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Ties",
                "description": "Only on the 409 from ending a show with a tied vote: the tied clips of every tied\ncategory, so the host breaks them all at once."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "Invite": {
        "type": "object",
        "required": [
          "email",
          "role",
          "createdAt"
        ],
        "properties": {
          "acceptedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "First sign-in."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "invitedBy": {
            "type": [
              "string",
              "null"
            ],
            "description": "Handle of the admin who sent it; empty for invites seeded from `ADMIN_EMAILS`."
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "Role the user gets on first sign-in."
          }
        }
      },
...
      "Me": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "required": [
              "shows",
              "showsForEveryone"
            ],
            "properties": {
              "shows": {
                "type": "boolean",
                "description": "The show (lobby, hosting, the hold switch on upload) is open to you."
              },
              "showsForEveryone": {
                "type": "boolean",
                "description": "...and to everyone, so uploads default to \"Save it for the show\"."
              }
            }
          }
        ],
        "description": "You, plus what's switched on for you."
      },
...
      "RevokeInvite": {
        "type": "object",
        "description": "Emails travel in the body, never the URL, so they stay out of HTTP logs.",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Role": {
        "type": "string",
        "enum": [
          "admin",
          "member"
        ]
      },
...
      "Ties": {
        "type": "object",
        "description": "The tied clips of each category when `end` needs the host's pick; empty for a\ncategory that isn't tied.",
        "required": [
          "clip",
          "fail"
        ],
        "properties": {
          "clip": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "fail": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "UpdateProfile": {
        "type": "object",
        "description": "Profile fields to change; omitted fields stay as they are.",
        "properties": {
          "displayName": {
            "type": [
              "string",
              "null"
            ],
            "description": "1–50 characters."
          },
          "handle": {
            "type": [
              "string",
              "null"
            ],
            "description": "2–32 characters: `a–z`, `0–9`, `-`, `_`, starting with a letter or digit."
          },
          "steamName": {
            "type": [
              "string",
              "null"
            ],
            "description": "Up to 64 characters; an empty string clears it."
          }
        },
        "additionalProperties": false
      },
      "UpdateUserAccess": {
        "type": "object",
        "properties": {
          "role": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/UserStatus",
                "description": "`disabled` blocks sign-in; their clips stay."
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "handle",
          "displayName",
          "role",
          "status",
          "createdAt"
        ],
        "properties": {
          "avatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "steamName": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}